Stop using radians wherever possible
It's a pain to constantly have to convert between radians and degrees,
and I couldn't keep track of which numbers were degrees and which were
radians. Wrapping the Python trigonometry functions is the easiest way
to avoid this -- the loss of precision due to floating-point
inaccuracies is negligible for our purposes.